// src/toolingLayer/pageObjects/common/artifacts/types.ts
// src/toolingLayer/pageObjects/common/artifacts/types.ts

export type PageArtifact = {
    pageKey: string;
    className: string;
    folderPath: string;
    elementsPath: string;
    aliasesGeneratedPath: string;
    aliasesHumanPath: string;
    pageObjectPath: string;
    registryImportPath: string;
};